refactor(ui): flatten `modals` dir
- `feedback` and `first-time-user` literally had a single file in them, so redundant dirs
- `new-version` had only two files in it

- interestingly, no `components/` dir here, so this is one example of where the previous directory structure was not internally consistent
  - there's a few others of those in the UI codebase

- rename `version.tsx` to `version.ts` as it has no JSX
  - rewrite it to a readable function instead of a large one-liner
  - remove deprecated / non-spec [`substr`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/substr) function and replace with `substring`
