+++ ui/src/app/modals/feedback-modal.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import {useState} from 'react';
-import {BigButton} from '../../shared/components/big-button';
-import {Modal} from '../../shared/components/modal/modal';
-import {SurveyButton} from '../../shared/components/survey-button';
+import {BigButton} from '../shared/components/big-button';
+import {Modal} from '../shared/components/modal/modal';
+import {SurveyButton} from '../shared/components/survey-button';
 
 export function FeedbackModal({dismiss}: {dismiss: () => void}) {
     const [choice, setChoice] = useState(0);
