+++ ui/src/app/modals/new-version-modal.scss
