+++ ui/src/app/modals/version.ts
@@ -1,0 +1,6 @@
+export function majorMinor(version: string) {
+    if (version.includes('.')) {
+        return version.substring(0, version.lastIndexOf('.') - 1);
+    }
+    return version;
+}
