--- ui/src/app/modals/version.tsx
@@ -1,1 +1,0 @@
-export const majorMinor = (version: string) => (version.includes('.') ? version.substr(0, version.lastIndexOf('.')) : version);
